+++ test/plugins/data/trigger/simple-normal-triggers.yaml
@@ -1,0 +1,28 @@
+shared:
+  image: node:20
+  steps:
+    - test: echo 'test'
+
+jobs:
+  a:
+    requires: [ ~commit, ~pr ]
+  b:
+    requires: [ a ]
+  c:
+    requires: [ a ]
+  d:
+    requires: [ b, c ]
+  e:
+    requires: [ ~d ]
+  f:
+    requires: [ ~d ]
+  g:
+    requires: [ ~e, ~f ]
+  h:
+    requires: [ ~g, g ]
+  i:
+    requires: [ ~g ]
+  j:
+    requires: [ ~g ]
+  k:
+    requires: [ ~h, i, j ]
