--- test/plugins/data/trigger/single-join.yaml
@@ -1,20 +1,0 @@
-shared:
-  image: node:20
-  steps:
-    - test: echo 'test'
-
-jobs:
-  first:
-    requires: [ ~commit, ~pr ]
-  second1:
-    requires: [ first ]
-  second2:
-    requires: [ first ]
-  third:
-    requires: [ second1, second2 ]
-  fourth:
-    requires: [ third, test ]
-  test:
-    requires: [ ~commit ]
-  detached:
-    requires: [ test, ~third ]
